// UniverCity-ui/src/app/Services/models/university-location.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { City } from '../models/city';
import { University } from '../models/university';
export interface UniversityLocation {
  city?: City;
  latitude?: number;
  longitude?: number;
  university?: University;
  universitylocation_id?: number;
}